backend/requirements.txt
# Wafrivet Field Vet — Backend Python dependencies
# Install with: pip install -r backend/requirements.txt

# --- Supabase Python client -------------------------------------------------
supabase>=2.9.0,<3.0.0

# --- Google Cloud Vertex AI (embeddings + Gemini Live) ----------------------
google-cloud-aiplatform>=1.75.0,<2.0.0
google-genai>=1.66.0,<2.0.0

# --- Google ADK (Phase 2+) --------------------------------------------------
google-adk>=1.26.0

# --- FastAPI WebSocket server (Phase 4+) ------------------------------------
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
websockets>=13.0,<16.0

# --- Phase 4: Session identity + direct Postgres (asyncpg + RLS) ------------
asyncpg>=0.29.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# --- Phase 5: PIN security + Redis pub/sub ----------------------------------
bcrypt>=4.1.0,<5.0.0
redis[hiredis]>=5.0.0,<6.0.0

# --- Environment and configuration ------------------------------------------
python-dotenv>=1.0.0,<2.0.0

# --- Logging and observability ----------------------------------------------
structlog>=24.0.0,<26.0.0

# --- Testing ----------------------------------------------------------------
pytest>=8.0.0,<10.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0
